yt graph simplification

+++ src/eez/gui/widgets/yt_graph.h
@@ -38,7 +38,7 @@ struct YTGraphWidgetState {
     WidgetState genericState;
     data::Value y2Data;
     uint8_t iChannel;
-    uint16_t historyValuePosition;
+    uint32_t historyValuePosition;
     uint16_t scanLinePosition;
     uint8_t ytGraphUpdateMethod;
 };
